split genetic_classes into separate files
